requirements.txt
# These requirements are for the Communication Subproblem
atproto 
geopy
google-auth
gspread
pprint
